sessions

+++ Server/src/index.js
@@ -5,7 +5,11 @@ const cors = require('cors');
 const {MongoClient, Collection, MongoAzureError} = require('mongodb');
 const api = require('./api.js');
 app.use(express.json())
-app.use(cors());
+app.use(cors({
+    origin: 'http://localhost:5000',
+    credentials: true,
+    methods: ['GET', 'POST', 'PUT', 'DELETE']
+}));
 const dburl = "mongodb+srv://victorlocherer:blQqG6A9ZpIX4p3Q@clusterprojet.etclz03.mongodb.net/"
 const client = new MongoClient(dburl);
 
@@ -24,10 +28,20 @@ app.use((req, res, next) => {
     next();
 });
 
+// middleware to test if authenticated
+function isAuthenticated (req, res, next) {
+    if (req.session.user) next()
+    else next('route')
+}
+
 app.use(session({
     secret: 'SuperProjet'
 }))
 app.post('/threads', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.CreateThread(req.db, req.body.original_poster_id, req.body.title, req.body.is_admin).then((thread_id) => {
         api.CreateServerMessage(req.db, thread_id.insertedId, req.body.original_poster_id, req.body.title, req.body.is_admin)
         res.status(200).json({thread_id: thread_id.insertedId});
@@ -38,6 +52,10 @@ app.post('/threads', async (req, res) => {
 });
 
 app.get('/threads/:thread_id', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.GetThreadMessages(req.db, req.params.thread_id).then((messages) => {
         res.status(200).json({messages: messages});
     }).catch(reason => {
@@ -46,6 +64,10 @@ app.get('/threads/:thread_id', async (req, res) => {
 });
 
 app.post('/threads/:thread_id', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.CreateMessage(req.db, req.params.thread_id, req.body.user_id, req.body.text).then(() => {
         res.status(200).json({message: "Message created"});
     }).catch(reason => {
@@ -54,6 +76,10 @@ app.post('/threads/:thread_id', async (req, res) => {
 });
 
 app.get('/users/:user_id', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     if (req.params.user_id === "0") {
         res.status(200).json({user: {username: "Server", logo: ""}, messages: []});
         return;
@@ -70,6 +96,10 @@ app.get('/users/:user_id', async (req, res) => {
 });
 
 app.put('/users/:user_id', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.PromoteUser(req.db, req.params.user_id).then(() => {
         res.status(200).json({message: "User promoted"});
     }).catch(reason => {
@@ -79,6 +109,10 @@ app.put('/users/:user_id', async (req, res) => {
 
 
 app.get('/threads', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.GetThreadByQuery(req.db, req.query.queryType, req.query.count).then((threads) => {
         res.status(200).json(threads);
     }).catch(reason => {
@@ -87,6 +121,14 @@ app.get('/threads', async (req, res) => {
 });
 
 app.delete('/users/:user_id', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
+    if(req.session.user.is_admin === false){
+        res.status(401).json({message: "User not admin"});
+        return;
+    }
     api.DeleteUser(req.db, req.params.user_id).then(() => {
         res.status(200).json({message: "User deleted"});
     }).catch(reason => {
@@ -119,12 +161,19 @@ app.post('/authentication/login', async (req, res) => {
 });
 
 app.get('/authentication/logout', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     req.session.destroy();
     res.status(200).json({message: "Logged out"});
 });
 
 app.get('/search', async (req, res) => {
-    console.log(req.query);
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.Search(req.db, req.query).then((results) => {
         res.status(200).json(results);
     }).catch(reason => {
@@ -133,6 +182,14 @@ app.get('/search', async (req, res) => {
 });
 
 app.post('/users/:user_id', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
+    if(req.session.user.is_admin === false){
+        res.status(401).json({message: "User not admin"});
+        return;
+    }
     api.ApproveUser(req.db, req.params.user_id).then(() => {
         res.status(200).json({message: "User approved"});
     }).catch(reason => {
@@ -141,6 +198,10 @@ app.post('/users/:user_id', async (req, res) => {
 });
 
 app.get('/users', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.GetUsersByQuery(req.db, req.query.queryType).then((users) => {
         res.status(200).json(users);
     }).catch(reason => {
@@ -149,6 +210,10 @@ app.get('/users', async (req, res) => {
 });
 
 app.post('/users', async (req, res) => {
+    if(!req.session.user){
+        res.status(401).json({message: "User not connected"});
+        return;
+    }
     api.CreateUser(req.db, req.body.login, req.body.password, req.body.admin).then((user_id) => {
         res.status(200).json({user_id: user_id.insertedId});
     })
